requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
langsmith>=0.1.75
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.30.0
anthropic>=0.28.0
